描画関数を分離
